requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
